+++ scripts/copy.js
@@ -1,0 +1,33 @@
+// const path = require('path');
+const fs      = require('fs');
+
+const mail    = 'https://github.com/LCluber/Type6.js';
+const CRLF    = '\r\n';
+const rel     = './';
+const src     = `${rel}build/`;
+const dest    = `${rel}dist/`; 
+const files   = [
+  {
+    src:  `${rel}src/type6.d.ts`,
+    dest: `${dest}type6.d.ts`
+  },
+  {
+    src:  `${src}type6.js`,
+    dest: `${dest}type6.js`
+  },
+];
+
+fs.mkdir(dest, { recursive: false },(err) => {
+  if (err) throw err;
+  fs.readFile(`${rel}LICENSE`, (err, license) => {
+    if (err) throw err;
+    for (let file of files) {
+      fs.readFile(file.src, (err, fileContent) => {
+        if (err) throw err;
+        fs.writeFile(file.dest, `/*${CRLF}${license}${CRLF}${mail}${CRLF}*/${CRLF}${CRLF}${fileContent}`, (err) => {
+          if (err) throw err;
+        });
+      });
+    }
+  });
+});